Move couch2sql stuff into StateDB
